fixing migration and insert data on database

+++ API/Entities/Photo.cs
@@ -5,11 +5,11 @@ namespace API.Entities
     [Table("Photos")]
     public class Photo
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Url { get; set; }
         public bool IsMain { get; set; }
         public string PublicId { get; set; }
         public AppUser AppUser { get; set; }
-        public int AppUserId { get; set; }
+        public long AppUserId { get; set; }
     }
 }
